# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "N100 Financial Platform API",
    "description": "REST API for the N100 Financial Platform - company fundamentals, screener, sector, peer, valuation and portfolio analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/peers/groups/{peer_group_name}": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Peer Group Members",
        "description": "List every company in a peer group, flagging the\nbenchmark company if one is set.\n\nRaises:\n    HTTPException: 404 if the peer group does not exist.",
        "operationId": "get_peer_group_members_api_v1_peers_groups__peer_group_name__get",
        "parameters": [
          {
            "name": "peer_group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Peer Group Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Peer Group Members Api V1 Peers Groups  Peer Group Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/peers/{ticker}/percentiles": {
      "get": {
        "tags": [
          "Peers"
        ],
        "summary": "Get Company Percentiles",
        "description": "Get every peer-percentile ranking recorded for a company.\n\nRaises:\n    HTTPException: 404 if no percentile data exists.",
        "operationId": "get_company_percentiles_api_v1_peers__ticker__percentiles_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Company Percentiles Api V1 Peers  Ticker  Percentiles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valuation/{ticker}": {
      "get": {
        "tags": [
          "Valuation"
        ],
        "summary": "Get Valuation History",
        "description": "Get the full market-cap / valuation history for a company.\n\nRaises:\n    HTTPException: 404 if the ticker has no valuation data.",
        "operationId": "get_valuation_history_api_v1_valuation__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Valuation History Api V1 Valuation  Ticker  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/valuation/rankings/top": {
      "get": {
        "tags": [
          "Valuation"
        ],
        "summary": "Get Valuation Rankings",
        "description": "Rank companies by their latest-year value for a given\nvaluation metric.\n\nRaises:\n    HTTPException: 400 if an unsupported metric is given.",
        "operationId": "get_valuation_rankings_api_v1_valuation_rankings_top_get",
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "One of: dividend_yield_pct, enterprise_value_crore, ev_ebitda, market_cap_crore, pb_ratio, pe_ratio",
              "default": "market_cap_crore",
              "title": "Metric"
            },
            "description": "One of: dividend_yield_pct, enterprise_value_crore, ev_ebitda, market_cap_crore, pb_ratio, pe_ratio"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Order"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Valuation Rankings Api V1 Valuation Rankings Top Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}